+++ src/features/writing/components/Modal.tsx
@@ -1,12 +1,12 @@
 import styled from 'styled-components';
 import { Icon } from '../../../components/ui/Icon';
-import Button from './Button';
 import useWritingResponseStore from '../../../stores/writingResponseStore';
 import { useMutation } from '@tanstack/react-query';
 import axios from 'axios';
 import { useNavigate } from 'react-router-dom';
 import { useUser } from '../../home/hooks/useUser';
 import showToast from '../utils/showToast';
+import { VisibilityDataTypes } from '../types/formTypes';
 
 interface ModalPropTypes {
   onClose: () => void;
@@ -15,19 +15,24 @@ interface ModalPropTypes {
   contentSecond: string;
 }
 
-interface VisibilityData {
-  user_id: string;
+interface ButtonPropTypes {
+  width?: string;
+  height?: string;
+  color?: string;
+  opacity?: number;
 }
 
 export default function Modal(props: ModalPropTypes) {
+  // 화면 이동 로직
+  const navigate = useNavigate();
+
+  // 사용자 정보 불러오는 로직
   const { getUserId } = useUser();
   const userId = getUserId().user_id;
 
-  const { contentId } = useWritingResponseStore((state) => state);
-  const navigate = useNavigate();
-
+  // 서버로 정보 전송 로직
   const mutation = useMutation({
-    mutationFn: async (userId: VisibilityData) => {
+    mutationFn: async (userId: VisibilityDataTypes) => {
       await axios.patch(`/api/diary/post/${contentId}/visibility`, userId);
     },
     onSuccess: () => {
@@ -42,6 +47,10 @@ export default function Modal(props: ModalPropTypes) {
     },
   });
 
+  // 클라이언트 정보 불러오는 로직
+  const { contentId } = useWritingResponseStore((state) => state);
+
+  // 클라이언트 정보 관리하는 로직
   const handleConfirm = () => {
     mutation.mutate({
       user_id: userId,
@@ -121,4 +130,16 @@ const ModalContent = styled.p`
   font-size: 1rem;
   color: #181a20;
   text-align: center;
+`;
+
+const Button = styled.button<ButtonPropTypes>`
+  width: ${({ width }) => width || '264px'};
+  height: ${({ height }) => height || '57px'};
+  background-color: ${({ color }) => color || '#9C99AE'};
+  opacity: ${({ opacity }) => (opacity !== undefined ? opacity : 1)};
+  border: none;
+  border-radius: 1rem;
+  color: #ffffff;
+  font-size: 16px;
+  cursor: pointer;
 `;
